requirements.txt
pyrogram
tgcrypto
pytgcalls
yt-dlp
youtube-search-python
flask
